Move rapidxml files to their own directory, add license
